requirements.txt
streamlit
pandas
openpyxl
dnspython